- reorg couple classes - added link from requests to storage
