+++ Datahub.Core/Model/Datahub/Datahub_Project.cs
@@ -293,51 +293,6 @@ namespace Datahub.Core.EFCore
         public Datahub_Project Project { get; set; }
     }
 
-    public class Datahub_ProjectServiceRequests
-    {
-        // TODO add requesting user to data model
-
-        [Key]
-        [AeFormIgnore]
-
-        public int ServiceRequests_ID { get; set; }
-
-        public DateTime ServiceRequests_Date_DT { get; set; }
-
-        public string ServiceType { get; set; }
-        public DateTime? Is_Completed { get; set; }
-        
-        [StringLength(200)]
-        public string User_Name { get; set; }
-
-        [StringLength(200)]
-        public string User_ID { get; set; }
-
-        public DateTime? Notification_Sent { get; set; }
-
-        [AeFormIgnore]
-        [Timestamp]
-        public byte[] Timestamp { get; set; }
-        public Datahub_Project Project { get; set; }
-    }
-
-    public class Datahub_Project_Costs
-    {
-        [Key]
-        public int ProjectCosts_ID { get; set; }
-
-        public int Project_ID { get; set; }
-
-        [StringLength(10)]
-        public string Project_Acronym_CD { get; set; }
-
-        public DateTime Usage_DT { get; set; }
-
-        public double Cost_AMT { get; set; }
-
-        public DateTime Updated_DT { get; set; }
-    }
-
     public class Datahub_Project_Sectors_And_Branches
     {
         [Key]
