+++ WebPortal/appsettings.sqlite.json
@@ -9,17 +9,18 @@
     "datahub-mssql-webanalytics": "Data Source=webanalytics.db",
     "datahub-cosmosdb": "Data Source=user-tracking.db",
     "datahub-mssql-metadata": "Data Source=metadata.db",
-    "datahub-mssql-languagetraining": "Data Source=languagetraining.db"
+    "datahub-mssql-languagetraining": "Data Source=languagetraining.db",
+    "datahub-mssql-m365forms": "Data Source=m365forms.db"
   },
   "AzureAd": {
     "Instance": "https://login.microsoftonline.com/",
-    "Domain": "163oxygen.onmicrosoft.com",
-    "TenantId": "8c1a4d93-d828-4d0e-9303-fd3bd611c822",
-    "ClientId": "b4d92f08-6ca8-4d71-90fd-d64c6c3ab5af",
-    "ClientSecret": "1R27Q~hRw-b__H5DLQiEpMn_aYGM5_gDJwbsH",
+    "Domain": "041gc.onmicrosoft.com",
+    "TenantId": "05c95b33-90ca-49d5-b644-288b930b912b",
+    "ClientId": "791c2276-28b6-4962-b043-381c47410892",
     "CallbackPath": "/signin-oidc",
+    "ClientSecret": "LrytFOTq5WlA/AgF5Lw75NO/G8TXFpm5ykqH1pvpoBM=",
     "SignedOutCallbackPath ": "/signout-callback-oidc",
-    "AppIDURL": "https://spdatahub-portal-app-sand.azurewebsites.net/"
+    "AppIDURL": "https://dh-portal-app-dev.azurewebsites.net/"
   },
   "Graph": {
     "BaseUrl": "https://graph.microsoft.com/v1.0",
